# contracts/trader-console/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "description": "Stable error envelope for resource failures.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "BarsResponse": {
        "additionalProperties": false,
        "description": "A page of OHLCV observations.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BarPoint"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "BarsResponse",
        "type": "object"
      },
...
      "ExperimentRunsResponse": {
        "additionalProperties": false,
        "description": "A page of runs belonging to one experiment.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "ExperimentRunsResponse",
        "type": "object"
      },
      "ExperimentSummary": {
        "additionalProperties": false,
        "description": "An experiment grouping discovered from its published run projections.",
        "properties": {
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "latest_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Created At"
          },
          "metadata_available": {
            "default": false,
            "title": "Metadata Available",
            "type": "boolean"
          },
          "run_count": {
            "title": "Run Count",
            "type": "integer"
          },
          "statuses": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Statuses",
            "type": "array"
          }
        },
        "required": [
          "experiment_id",
          "run_count"
        ],
        "title": "ExperimentSummary",
        "type": "object"
      },
      "ExperimentsResponse": {
        "additionalProperties": false,
        "description": "A page of experiment groups.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "ExperimentsResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketDatasetsResponse": {
        "additionalProperties": false,
        "description": "Available market-data slices with bounded pagination.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MarketDataset"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "MarketDatasetsResponse",
        "type": "object"
      },
      "PageInfo": {
        "additionalProperties": false,
        "description": "Bounded pagination evidence returned with collection resources.",
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "offset",
          "total",
          "has_more"
        ],
        "title": "PageInfo",
        "type": "object"
      },
...
      "RiskDecisionsResponse": {
        "additionalProperties": false,
        "description": "A bounded, filterable page of manager decisions for one run.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RiskDecision"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "RiskDecisionsResponse",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "description": "Published evidence for one backtest run, grouped by projection.",
        "properties": {
          "assumptions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "comparison_curves": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Comparison Curves",
            "type": "array"
          },
          "comparison_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "equity_curve": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Equity Curve",
            "type": "array"
          },
          "evidence_coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "exposure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "fills": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Fills",
            "type": "array"
          },
          "indicator_series": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IndicatorSeriesPoint"
            },
            "title": "Indicator Series",
            "type": "array"
          },
          "orders": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Orders",
            "type": "array"
          },
          "performance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "positions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Positions",
            "type": "array"
          },
          "provenance": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Provenance",
            "type": "array"
          },
          "risk_composition": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RiskCompositionEntry"
            },
            "title": "Risk Composition",
            "type": "array"
          },
          "risk_decisions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RiskDecision"
            },
            "title": "Risk Decisions",
            "type": "array"
          },
          "risk_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "run": {
            "$ref": "#/components/schemas/ExperimentRunSummary"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "signal_markers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SignalMarker"
            },
            "title": "Signal Markers",
            "type": "array"
          },
          "signals": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Signals",
            "type": "array"
          },
          "trades": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Trades",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Trader Console API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/experiments": {
      "get": {
        "description": "Discover experiment groups from published backtest runs.",
        "operationId": "experiments_api_experiments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Experiments",
        "tags": [
          "resources"
        ]
      }
    },
...
    "/api/experiments/{experiment_id}/runs": {
      "get": {
        "description": "List runs and comparison eligibility within one experiment.",
        "operationId": "experiment_runs_api_experiments__experiment_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "compatible_with_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compatible With Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Experiment Runs",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/market-data/bars": {
      "get": {
        "description": "Return bounded OHLCV observations ordered from earliest to latest.",
        "operationId": "market_data_bars_api_market_data_bars_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "asset_class",
            "required": false,
            "schema": {
              "default": "stock",
              "enum": [
                "stock",
                "crypto"
              ],
              "title": "Asset Class",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "1Min",
              "maxLength": 32,
              "minLength": 1,
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 50000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Market Data Bars",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/market-data/datasets": {
      "get": {
        "description": "List available symbol/timeframe/source market-data slices.",
        "operationId": "market_datasets_api_market_data_datasets_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDatasetsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Market Datasets",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "description": "Return a backtest run and its published evidence sections.",
        "operationId": "run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "section_limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Section Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run Detail",
        "tags": [
          "resources"
        ]
      }
    },
    "/api/runs/{run_id}/risk-decisions": {
      "get": {
        "description": "Return a bounded, filterable manager decision trace for one run.",
        "operationId": "risk_decisions_api_runs__run_id__risk_decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "manager_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manager Id"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "approved",
                    "transformed",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "cycle_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cycle Id"
            }
          },
          {
            "in": "query",
            "name": "client_order_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Order Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDecisionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Risk Decisions",
        "tags": [
          "resources"
        ]
      }
    },
...
  }
}